% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Clustering}
\alias{Clustering}
\title{Clustering analysis}
\usage{
Clustering(dat, method = c("PCA","RPCA"))
}
\arguments{
\item{x}{A matrix or dataframe with objects in the rows and features in the columns}

\item{method}{a string with one of the following values:  \code{PCA}, \code{RPCA}}

\item{scale}{a boolean indicating if the variables should be scaled.}
}
\value{
plot object
}
\description{
Clustering analysis
}
\examples{
require(mvtnorm)
 data(iris)
 Clustering(iris[,-5],method="PCA")
}
\author{
Rafael S. de Souza
}
\keyword{misc}

