% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Corr_MIC.R
\name{Corr_MIC}
\alias{Corr_MIC}
\title{Non-linear correlation}
\usage{
Corr_MIC(x, method = c("pearson","spearman", "MIC"), scale = "")
}
\arguments{
\item{x}{A matrix or dataframe with objects in the rows and features in the columns}

\item{method}{a string with one of the following values:  \code{pearson}, \code{spearman},\code{MIC}}

\item{scale}{a boolean indicating if the variables should be scaled.}
}
\value{
Correlation Matrix
}
\description{
Non-linear correlation
}
\details{
The program is a simple alteration of cor() and MIC from package Minerva.
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
}
\author{
Rafael S. de Souza
}
\keyword{misc}

