% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Nightingale.R
\name{Nightingale}
\alias{Nightingale}
\title{PCA  visualization of a hierarchical clustering from a correlation matrix}
\usage{
Nightingale(corr, npcs= 4,PCAmethod = c("PCA","RPCA"))
}
\arguments{
\item{corr}{Correlation Matrix}

\item{npcs}{number of principal components}

\item{PCAmethod}{a string with one of the following values:  \code{PCA}, \code{RPCA}}
}
\value{
ggplot object
}
\description{
\code{Nightingale} computes Robust PCA projections of
variables and plot them in polar coordinates.
}
\details{
The program is a simple alteration of PCAproj() that computes a desired number of robust
 principal components using the grid search algorithm in a sphere.
}
\examples{
data(iris)
 require(mvtnorm)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 Nightingale(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

