% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotchord}
\alias{plotchord}
\title{Plot a chord diagram from a correlation matrix}
\usage{
plotchord(x,color="RdBu",...)
}
\arguments{
\item{x}{a correlation  matrix}

\item{color}{a list with color pallets: "RdBu","YlGn","PRGn","OrRd", default = "RdBu"}

\item{...}{other options for distogram function}
}
\value{
A plot of the Chord diagram
}
\description{
Plot a chord diagram from a correlation matrix
}
\details{
The program is a simple wrap  of chordDiagram() from package circlize.
}
\examples{
data(iris)
cor1<-Corr_MIC(iris[,1:4],method="pearson")
plotchord(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

