% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotgraph.R
\name{plotgraph}
\alias{plotgraph}
\title{Network visualization of a hierarchical clustering from a correlation matrix}
\usage{
plotgraph(corr,layout=c("spring","circular"))
}
\arguments{
\item{corr}{Correlation Matrix}

\item{layout}{string}
}
\value{
qgraph object
}
\description{
Network visualization of a hierarchical clustering from a correlation matrix
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 plotgraph(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

