% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotheatmap}
\alias{plotheatmap}
\title{Plot a heatmap  from a correlation matrix}
\usage{
plotheatmap(x,display_numbers=FALSE,...)
}
\arguments{
\item{x}{a correlation  matrix}

\item{display_numbers}{logical, default = FALSE}

\item{...}{other options for  pheatmap function}
}
\value{
A plot of the heatmap
}
\description{
Plot a heatmap  from a correlation matrix
}
\details{
The program is a simple wrap  of pheatmap() from package pheatmap.
}
\examples{
data(iris)
cor1<-Corr_MIC(iris[,1:4],method="pearson")
plotheatmap(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

