% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sunburstPCA}
\alias{sunburstPCA}
\title{PCA  visualization of a hierarchical clustering from a correlation matrix}
\usage{
sunburstPCA(corr, npcs= 4)
}
\arguments{
\item{corr}{Correlation Matrix}

\item{npcs}{number of principal components}
}
\value{
ggplot object
}
\description{
\code{sunburstPCA} computes Robust PCA projections of
variables and plot them in polar coordinates.
}
\details{
The program is a simple alteration of PCAproj() that computes a desired number of robust
 principal components using the grid search algorithm in a sphere.
}
\examples{
data(iris)
 require(mvtnorm)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 sunburstPCA(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

