% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotcorrDist}
\alias{plotcorrDist}
\title{Plot a distogram fom a correlation matrix}
\usage{
plotcorrDist(x,labels=NULL,...)
}
\arguments{
\item{x}{a correlation  matrix}

\item{labels}{a list with variables names, default = NULL}

\item{...}{other options for distogram function}
}
\value{
A plot of the Correlation Matrix
}
\description{
Plot a distogram fom a correlation matrix
}
\details{
The program is a simple alteration of distogram().
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 plotcorrDist(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

