% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotdendrogram}
\alias{plotdendrogram}
\title{Plot a dendrogram  from a correlation matrix}
\usage{
plotdendrogram(corr,type=c("phylogram", "cladogram", "fan"))
}
\arguments{
\item{corr}{Correlation Matrix}

\item{type}{logical}
}
\value{
Dendrogram colored by clusters
}
\description{
Plot a dendrogram  from a correlation matrix
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 plotdendrogram(cor1,type="p")
}
\author{
Rafael S. de Souza
}
\keyword{misc}

