% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotdendrogram}
\alias{plotdendrogram}
\title{Dendrogram of a hierarchical clustering from a correlation matrix}
\usage{
plotdendrogram(corr)
}
\arguments{
\item{corr}{Correlation Matrix}
}
\value{
Dendrogram colored by clusters
}
\description{
Dendrogram of a hierarchical clustering from a correlation matrix
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 plotdendrogram(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

