% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotgraph}
\alias{plotgraph}
\title{Plot a network visualization  from a correlation matrix}
\usage{
plotgraph(corr,layout=c("spring","circular"))
}
\arguments{
\item{corr}{Correlation Matrix}

\item{layout}{string}
}
\value{
qgraph object
}
\description{
Plot a network visualization  from a correlation matrix
}
\examples{
data(iris)
 cor1<-Corr_MIC(iris[,1:4],method="pearson")
 plotgraph(cor1)
}
\author{
Rafael S. de Souza
}
\keyword{misc}

